Improve cache for architectures (#26)
* Add runner.arch for cache.key

Users might use this action to prepare multiple jobs
with the same OS but different architectures, for example,
macOS' x86_64 and arm64, Linux's x86_64 and arm64 and even
Windows x86_64 and arm64.

This CL just add architecture information of runner
to make cache workable for this special scenario.

Signed-off-by: utzcoz <utzcoz@outlook.com>

* ci: Add macos-14, macos + arm64 runner target

It can be used to test cache between same OS but different
architectures.

This CL also adds a small step to print runner's
os and arch information for debug purpose.

Signed-off-by: utzcoz <utzcoz@outlook.com>

* ci: Enable pull_request on main branch to trigger CI for PR

Signed-off-by: utzcoz <utzcoz@outlook.com>

---------

Signed-off-by: utzcoz <utzcoz@outlook.com>